TASK-86 [Feat]:사이드바 컴포넌트 세분화 (1차 리팩토링 완료)
