+++ src/components/@shared/Common/SideBar.tsx
@@ -1,166 +1,18 @@
-'use client';
-
-import { MouseEvent, PropsWithChildren, useCallback, useState } from 'react';
-import { SubmitHandler, useForm } from 'react-hook-form';
-
-import { Flex, Modal, Stack, Text, UnstyledButton } from '@mantine/core';
+import { Flex, Stack, Text, UnstyledButton } from '@mantine/core';
 import { useDisclosure } from '@mantine/hooks';
 import Image from 'next/image';
 import Link from 'next/link';
-import { useRouter } from 'next/navigation';
 
-import postCreateDashboard from '@core/api/postDashboard';
-import { useRoot } from '@core/contexts/RootContexts';
-import COLORS from '@lib/constants/themeConst';
-import useDashBoards from '@lib/hooks/useDashBoards';
-import useDevice, { DEVICE } from '@lib/hooks/useDevice';
-import cn from '@lib/utils/cn';
+import useDevice from '@lib/hooks/useDevice';
 
-import PrimaryButton from '../UI/Button/PrimaryButton';
-import SecondaryButton from '../UI/Button/SecondaryButton';
-
-import Input from './Input';
-
-type ColorType = keyof typeof COLORS;
-
-interface Inputs {
-  title: string;
-}
-
-type DeviceKeyObject = {
-  [key in keyof typeof DEVICE]: string;
-};
-
-const MODAL_SIZE: DeviceKeyObject = {
-  mobile: '327px',
-  tablet: '584px',
-  desktop: '584px',
-};
-
-const MODAL_RADIUS: DeviceKeyObject = {
-  mobile: '8px',
-  tablet: '16px',
-  desktop: '16px',
-};
-
-const DashBoardAddModal = ({
-  children,
-  opened,
-  onClose,
-  redirect,
-}: PropsWithChildren<{
-  opened: boolean;
-  onClose: () => void;
-  redirect: (id: number) => void;
-}>) => {
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm<Inputs>({
-    defaultValues: {
-      title: '',
-    },
-  });
-  const device = useDevice();
-  const [color, setColor] = useState<ColorType>(COLORS.green);
-
-  const onSubmit: SubmitHandler<Inputs> = async ({ title }) => {
-    let res;
-    try {
-      res = await postCreateDashboard({ title, color });
-      redirect(res.data.id);
-    } catch {
-      // eslint-disable-next-line no-console
-      console.log('네트워크 에러가 발생하였습니다. 잠시 후 다시 시도해주세요');
-    }
-  };
-
-  const handleColorClick = (e: MouseEvent<HTMLButtonElement>) => {
-    const { target } = e;
-    if (!(target instanceof HTMLElement)) return;
-    const { color: colorSet } = target.dataset;
-    setColor(colorSet as ColorType);
-  };
-
-  return (
-    <>
-      <Modal
-        size={MODAL_SIZE[device]}
-        radius={MODAL_RADIUS[device]}
-        opened={opened}
-        onClose={onClose}
-        withCloseButton={false}
-        centered
-      >
-        <Stack className="gap-4 md:p-4" onSubmit={handleSubmit(onSubmit)}>
-          <h2 className="font-xl-20px-bold">새로운 대시보드</h2>
-          <Input
-            className="border border-border-gray"
-            id="title"
-            label="대시보드 이름"
-            type="text"
-            placeholder="뉴프로젝트"
-            register={register}
-            errors={errors}
-            validation={{
-              required: '대시보드 이름을 입력해주세요',
-            }}
-          />
-          <Flex className="gap-2">
-            {Object.keys(COLORS).map(colorKey => (
-              <button
-                className="flex h-[30px] w-[30px] items-center justify-center rounded-full"
-                style={{ backgroundColor: COLORS[colorKey] }}
-                key={colorKey}
-                data-color={COLORS[colorKey]}
-                onClick={handleColorClick}
-                type="button"
-              >
-                {color === COLORS[colorKey] && (
-                  <Image
-                    width={24}
-                    height={24}
-                    src="/icons/check.png"
-                    alt="color check"
-                  />
-                )}
-              </button>
-            ))}
-          </Flex>
-          <Flex className="mt-4 gap-2">
-            <SecondaryButton onClick={onClose}>취소</SecondaryButton>
-            <PrimaryButton onClick={handleSubmit(onSubmit)}>확인</PrimaryButton>
-          </Flex>
-        </Stack>
-      </Modal>
-      {children}
-    </>
-  );
-};
+import DashBoardAddModal from './Modals/DashBoardAddModal';
+import SideBarList from './SideBarList';
 
 export default function SideBar() {
-  const router = useRouter();
-  const { dashboardid, setDashboardid } = useRoot();
-  const { dashBoardList } = useDashBoards();
   const device = useDevice();
   const [opened, { open, close }] = useDisclosure(false);
 
   const isMobile = device === 'mobile';
-
-  const { dashboards } = dashBoardList;
-
-  const redirectDashboard = useCallback(
-    (id: number) => {
-      setDashboardid(String(id));
-      router.push(`/dashboard/${id}`);
-    },
-    [router, setDashboardid]
-  );
-
-  const handleDashboardClick = (id: number) => {
-    redirectDashboard(id);
-  };
 
   return (
     <Stack className="fixed bottom-0 left-0 top-0 z-50 w-[67px] items-center border-r border-border-gray bg-white pt-5 text-gray-400 md:w-40 md:items-stretch md:px-[13px] xl:w-[300px]">
@@ -182,11 +34,7 @@ export default function SideBar() {
           />
         )}
       </Link>
-      <DashBoardAddModal
-        opened={opened}
-        onClose={close}
-        redirect={redirectDashboard}
-      >
+      <DashBoardAddModal opened={opened} onClose={close}>
         <Flex className="mt-4 items-center justify-between md:mt-10">
           {!isMobile && (
             <Text className="font-xs-12px-semibold">Dash Boards</Text>
@@ -199,43 +47,7 @@ export default function SideBar() {
           </UnstyledButton>
         </Flex>
       </DashBoardAddModal>
-      <Stack className="items-center gap-2 md:items-stretch">
-        {dashboards &&
-          dashboards?.map(dashboard => (
-            <Flex
-              key={dashboard.id}
-              className={cn(
-                'h-[43px] items-center rounded hover:border-2 hover:border-blue',
-                dashboard.id === Number(dashboardid) && 'bg-violet-white'
-              )}
-              onClick={() => handleDashboardClick(dashboard.id)}
-            >
-              <button className="flex h-10 w-10 items-center justify-center">
-                <div
-                  className="h-2 w-2 rounded-full"
-                  style={{ backgroundColor: dashboard.color }}
-                  aria-label="link button"
-                />
-              </button>
-              {!isMobile && (
-                <Flex className="items-center gap-2">
-                  <span className="h-[26px] max-w-[75px] overflow-hidden text-ellipsis whitespace-nowrap font-lg-16px-medium xl:max-w-[200px] xl:font-2lg-18px-medium">
-                    {dashboard.title}
-                  </span>
-                  {dashboard.createdByMe && (
-                    <Image
-                      width={16}
-                      height={12}
-                      src="/icons/crown.png"
-                      alt="createdByMe"
-                    />
-                  )}
-                </Flex>
-              )}
-            </Flex>
-          ))}
-      </Stack>
-      <Stack />
+      <SideBarList />
     </Stack>
   );
 }
